Base: huge cleanup
*) catch unregistering receivers to prevent crashes
*) remove unused imports
*) remove unneeded semicolons
*) manual array copy
*) string comparisons with == instead of equals()
*) unneccessary boxing and unboxing
*) codestyle
*) update gradle version
*) much much more

+++ core/java/com/android/internal/widget/LockPatternView.java
@@ -1138,7 +1138,7 @@ public class LockPatternView extends View {
             super(in);
             mSerializedPattern = in.readString();
             mDisplayMode = in.readInt();
-            mPatternSize = (byte) in.readByte();
+            mPatternSize = in.readByte();
             mInputEnabled = (Boolean) in.readValue(null);
             mInStealthMode = (Boolean) in.readValue(null);
             mTactileFeedbackEnabled = (Boolean) in.readValue(null);
